Support limiting symbol deletion to certain paths
Summary:
For Nefarious Facebook Internal Purposes, which may or may not
include incremental symbol database updates.

Give the import script an option to not clear all symbols from the
project, and make a new script that clears only symbols from paths given
on stdin.

(I'm not yet sure how much of the NFIPs is going to be ported over
here. So there might be a future diff that uses this. Conversely,
since there's no real use case out here, I'm fine just moving this to
the Facebook configuration if necessary.)

Test Plan: Run scripts in innumerable bizarre and eldritch configurations.

Reviewers: epriestley

Reviewed By: epriestley

CC: aran, Korvin

Maniphest Tasks: T1602

Differential Revision: https://secure.phabricator.com/D3305

+++ scripts/symbols/import_project_symbols.php
@@ -31,6 +31,11 @@ $args->parseStandardArguments();
 $args->parse(
   array(
     array(
+      'name'      => 'no-purge',
+      'help'      => 'Do not clear all symbols for this project before '.
+                     'uploading new symbols. Useful for incremental updating.',
+    ),
+    array(
       'name'      => 'more',
       'wildcard'  => true,
     ),
@@ -151,12 +156,14 @@ foreach ($symbols as $dict) {
     $path_map[$dict['path']]);
 }
 
-echo "Purging old symbols...\n";
-queryfx(
-  $conn_w,
-  'DELETE FROM %T WHERE arcanistProjectID = %d',
-  $symbol->getTableName(),
-  $project->getID());
+if (!$args->getArg('no-purge')) {
+  echo "Purging old symbols...\n";
+  queryfx(
+    $conn_w,
+    'DELETE FROM %T WHERE arcanistProjectID = %d',
+    $symbol->getTableName(),
+    $project->getID());
+}
 
 echo "Loading ".number_format(count($sql))." symbols...\n";
 foreach (array_chunk($sql, 128) as $chunk) {
